EIP-92: Adding requested changes
